Fix installer failing to patch AddIn file due to access rights (#16)
Rather than patch the AddIn file in the installation directory (C:\Program Files...), save the patched version to the user's temp directory, then install this temporary patched version. On Windows 10, even when using an administrator account with UAC set to "never notify", the installer will fail because the post-install hook "OnInstall" does not run with the same elevated rights as the MSI does. This change just makes it so that this post-install hook patches the AddIn file in the user's temporary folder rather than the install directory.