Switched the VSIX package class to derive from AsyncPackage and do async init.
Also removing the old VS addin as well as support to install the VSIX in anything but 2017 or 2019. I don't use older versions of VS anymore and can't test on these versions. Only 2015 or up, I believe, support async package.

Note that the command line RudeBuildConsole exe still fully supports older VS versions. But if you want to install the VS extension you need some version of 2017 or 2019, which given that the Community edition is free shouldn't be too hard of a burden.