Add support to launch MSBuild directly to build the RudeBuild-generated .sln/.vcxproj files. This is now the default as it's faster than launching devenv to build the .sln.
To allow for building projects that are located in a solution folder in the .sln needed to parse the nested folders section out of the .sln and keep track of the solution folders in SolutionInfo and the SolutionConfigManager to associated projects with the solution folder they reside in.
That's because when specifying a .sln on the MSBuild command line and only wanting to build a specific project the -target command line argument needs to be used and that requires specifying the fully qualified path (in terms of solution folders) to the project.

Also added a FileNameSuffix in addition to the existing prefix to the GlobalSettings. Changed the default to use a .RudeBuild suffix instead of the RudeBuild_ prefix because that keeps the generated .sln/.vcxproj right by the original .sln/.vcxproj files in alphabetical sort order in Explorer.